% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db.R
\name{connect_db}
\alias{connect_db}
\title{Connect to database}
\usage{
connect_db(drv = c("MySQL", "SQLite"), setup = FALSE, testuser = FALSE, ...)
}
\arguments{
\item{drv}{Database driver, either SQLite or MySQL}

\item{setup}{Setup database schema, TRUE/FALSE}

\item{testuser}{Create a test user, TRUE/FALSE}

\item{...}{Other parameters passed to \code{\link{connect_mysql}}}
}
\value{
A database connection object
}
\description{
Connect to database
}
\examples{
\dontrun{
connect_db("MySQL", setup = FALSE, testuser = FALSE)
connect_db("SQLite", setup = TRUE, testuser = TRUE)
connect_db("MySQL", setup = TRUE, testuser = TRUE)
}
}
