% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_mysql.R
\name{connect_mysql}
\alias{connect_mysql}
\title{Title}
\usage{
connect_mysql(
  dbname = Sys.getenv("MYSQL_ADDON_DB"),
  host = Sys.getenv("MYSQL_ADDON_HOST"),
  port = Sys.getenv("MYSQL_ADDON_PORT"),
  user = Sys.getenv("MYSQL_ADDON_USER"),
  password = Sys.getenv("MYSQL_ADDON_PASSWORD")
)
}
\arguments{
\item{dbname}{}

\item{host}{}

\item{port}{}

\item{user}{}

\item{password}{}
}
\description{
Title
}
\examples{
db <- connect_mysql()
DBI::dbDisconnect(db)
rm(db)
}
