% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_sqlite.R
\name{connect_sqlite}
\alias{connect_sqlite}
\title{Connect SQLite}
\usage{
connect_sqlite()
}
\value{
A SQLite database connection object
}
\description{
Connect to a temporary SQLite database stored in memory
}
\examples{
\dontrun{connect_sqlite()}
}
