% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_users.R
\name{create_dummy_user}
\alias{create_dummy_user}
\title{Create a dummy user}
\usage{
create_dummy_user(db)
}
\arguments{
\item{db}{Database connection object.}
}
\description{
Create dummy user
}
\examples{
\dontrun{
db <- connect_sqlite()
setup_db_schema(db)
create_dummy_user(db)
}
}
