% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loginui.R
\name{loginUI}
\alias{loginUI}
\title{Shiny Login Module}
\usage{
loginUI(id)
}
\arguments{
\item{id}{Namespace id}
}
\value{
A reactive values data frame with the following fields: \itemize{
\item{verified}{ Boolean, whether or not the user is verified.}
\item{message}{ A character string with any message from server to user.}
}
}
\description{
Verify user credentials against authentication database.
}
\examples{
\dontrun{
library(shiny)
library(rlogin)

ui <- fluidPage(
  loginUI("login")
)

server <- function(input, output, session) {

  callModule(
    module = loginServer,
    id = "login"
  )
}

shinyApp(ui, server)
}
}
