% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyLogin.R
\name{rlogin}
\alias{rlogin}
\alias{loginUI}
\alias{logoutUI}
\alias{loginServer}
\title{Shiny Login Module}
\usage{
loginUI(id)

logoutUI(id)

loginServer(input, output, session, db)
}
\arguments{
\item{id}{Namespace id}

\item{db}{Database connection object}
}
\value{
A reactive values data frame with the following fields: \itemize{
\item{verified}{ Boolean, whether or not the user is verified.}
\item{message}{ A character string with any message from server to user.}
}
}
\description{
Verify user credentials against authentication database.
}
\section{Functions}{
\itemize{
\item \code{loginUI()}: loginUI

\item \code{logoutUI()}: logoutUI

\item \code{loginServer()}: loginServer

}}
\examples{
\dontrun{
library(shiny)
library(rlogin)

ui <- fluidPage(
  loginUI("login")
)

server <- function(input, output, session) {

  callModule(
    module = loginServer,
    id = "login"
  )
}

shinyApp(ui, server)
}
}
