% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_auth_cookie.R
\name{set_auth_cookie}
\alias{set_auth_cookie}
\title{Set Authentication Cookie}
\usage{
set_auth_cookie(db, userid)
}
\arguments{
\item{db}{Database connection}

\item{userid}{User id}
}
\description{
Set an authentication cookie and store values in database for lookup
}
\examples{
\dontrun{set_auth_cookie(db, userid)}
}
