% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-mysql-schema.R
\name{setup_db_schema}
\alias{setup_db_schema}
\title{Setup db schema}
\usage{
setup_db_schema(
  dbname = Sys.getenv("MYSQL_ADDON_DB"),
  host = Sys.getenv("MYSQL_ADDON_HOST"),
  port = Sys.getenv("MYSQL_ADDON_PORT"),
  user = Sys.getenv("MYSQL_ADDON_USER"),
  password = Sys.getenv("MYSQL_ADDON_PASSWORD"),
  interactive = TRUE
)
}
\arguments{
\item{dbname}{Database name}

\item{host}{Database host}

\item{port}{Database post}

\item{user}{Database username}

\item{password}{Database password}
}
\description{
Create two tables in a MySQL database called auth_tokens, and auth_user
}
\examples{
\dontrun{
setup_db_schema()
}
}
