% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_password.R
\name{verify_password}
\alias{verify_password}
\title{Verify Password}
\usage{
verify_password(
  input_username = input$username,
  input_password = input$password,
  dbname = Sys.getenv("MYSQL_ADDON_DB"),
  host = Sys.getenv("MYSQL_ADDON_HOST"),
  port = Sys.getenv("MYSQL_ADDON_PORT"),
  user = Sys.getenv("MYSQL_ADDON_USER"),
  password = Sys.getenv("MYSQL_ADDON_PASSWORD")
)
}
\arguments{
\item{input_username}{User inputted user name.}

\item{input_password}{User inputted password.}

\item{dbname}{Database name}

\item{host}{Database host}

\item{port}{Database post}

\item{user}{Database username}

\item{password}{Database password}
}
\value{
Logical value of TRUE or FALSE depending on whether the password is verified.
}
\description{
Function to verify hashed password.
}
\examples{
\dontrun{
verify_password()
}
}
