% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_mysql.R
\name{connect_mysql}
\alias{connect_mysql}
\title{Connect MySQL}
\usage{
connect_mysql(
  dbname = Sys.getenv("MYSQL_ADDON_DB"),
  host = Sys.getenv("MYSQL_ADDON_HOST"),
  port = Sys.getenv("MYSQL_ADDON_PORT"),
  user = Sys.getenv("MYSQL_ADDON_USER"),
  password = Sys.getenv("MYSQL_ADDON_PASSWORD")
)
}
\arguments{
\item{dbname}{Database name}

\item{host}{Database host}

\item{port}{Database post}

\item{user}{Database username}

\item{password}{Database password}
}
\value{
A database connection object
}
\description{
Connect to a MySQL database
}
\examples{
\dontrun{
  db <- connect_mysql()
}
}
