% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-mysql-schema.R
\name{create_dummy_user}
\alias{create_dummy_user}
\title{Create a dummy user}
\usage{
create_dummy_user(
  dbname = Sys.getenv("MYSQL_ADDON_DB"),
  host = Sys.getenv("MYSQL_ADDON_HOST"),
  port = Sys.getenv("MYSQL_ADDON_PORT"),
  user = Sys.getenv("MYSQL_ADDON_USER"),
  password = Sys.getenv("MYSQL_ADDON_PASSWORD")
)
}
\arguments{
\item{dbname}{Database name}

\item{host}{Database host}

\item{port}{Database post}

\item{user}{Database username}

\item{password}{Database password}
}
\description{
Create dummy user
}
\examples{
\dontrun{
create_dummy_user()
}
}
