% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_auth_cookie.R
\name{get_auth_token}
\alias{get_auth_token}
\title{Get Authentication Token}
\usage{
get_auth_token(db, token)
}
\arguments{
\item{db}{Database connection object}
}
\value{
If authenticated then return verified userid
}
\description{
Get token and cross reference against stored data
}
\examples{
\dontrun{get_auth_token(db)}
}
