% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-mysql-schema.R
\name{setup_db_schema}
\alias{setup_db_schema}
\title{Setup db schema}
\usage{
setup_db_schema(db = connect_sqlite(), interactive = TRUE)
}
\arguments{
\item{db}{Database connection from \code{\link{connect_mysql}} or \code{\link{connect_sqlite}}.}

\item{interactive}{Boolean TRUE/FALSE variable. Specifies whether in interactive session.}
}
\description{
Create two tables in a MySQL database called auth_tokens, and auth_user
}
\examples{
\dontrun{
db <- connect_sqlite()
setup_db_schema(db)
}
}
