% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_password.R
\name{verify_password}
\alias{verify_password}
\title{Verify Password}
\usage{
verify_password(db, input_username, input_password)
}
\arguments{
\item{db}{Database connection object}

\item{input_username}{User inputted user name.}

\item{input_password}{User inputted password.}
}
\value{
Logical value of TRUE or FALSE depending on whether the password is verified.
}
\description{
Function to verify hashed password.
}
\examples{
\dontrun{
verify_password(db, input$username, input$password)
}
}
